Added helper class to consolidate boilerplate setup of tests for diagnostic entries
