+++ identity-server/test/IdentityServer.UnitTests/Licensing/v2/DiagnosticEntries/AssemblyInfoDiagnosticEntryTests.cs
@@ -1,8 +1,6 @@
 // Copyright (c) Duende Software. All rights reserved.
 // See LICENSE in the project root for license information.
 
-using System.Buffers;
-using System.Text;
 using System.Text.Json;
 using Duende.IdentityServer.Licensing.V2.Diagnostics.DiagnosticEntries;
 
@@ -14,17 +12,10 @@ public class AssemblyInfoDiagnosticEntryTests
     public async Task WriteAsync_ShouldWriteAssemblyInfo()
     {
         var subject = new AssemblyInfoDiagnosticEntry();
-        var bufferWriter = new ArrayBufferWriter<byte>();
-        await using var writer = new Utf8JsonWriter(bufferWriter, new JsonWriterOptions { Indented = false });
-        writer.WriteStartObject();
 
-        await subject.WriteAsync(writer);
+        var result = await DiagnosticEntryTestHelper.WriteEntryToJson(subject);
 
-        writer.WriteEndObject();
-        await writer.FlushAsync();
-        var json = Encoding.UTF8.GetString(bufferWriter.WrittenSpan);
-        var jsonDocument = JsonDocument.Parse(json);
-        var assemblyInfo = jsonDocument.RootElement.GetProperty("AssemblyInfo");
+        var assemblyInfo = result.RootElement.GetProperty("AssemblyInfo");
         assemblyInfo.GetProperty("AssemblyCount").ValueKind.ShouldBe(JsonValueKind.Number);
         var assemblies = assemblyInfo.GetProperty("Assemblies");
         assemblies.ValueKind.ShouldBe(JsonValueKind.Array);
